Base Functions

Plotting Functions

QuantumAnnealingAnalytics.plot_annealing_schedule — Method

function to plot an annealing schedule from QuantumAnnealing.jl. kwargs are for Plots.plot

QuantumAnnealingAnalytics.plot_ground_states_dwisc — Method

function to plot only the lowest energy states in a dwisc bqpjson format file, String, or Dict. possible orderings are :numeric or :hamming. spin_comp specifies a state in spin format for comparison in hamming distance calculations. kwargs is for Plots.bar

QuantumAnnealingAnalytics.plot_state_steps — Method

function to plot the state steps, generated from calling simulate(..., state_steps=[]). This is used to see instantaneous measurement values throughout the anneal. kwargs are for Plots.plot

QuantumAnnealingAnalytics.plot_states — Method

function to plot the states present in a density matrix output by QuantumAnnealing.simulate kwargs are for Plots.bar

QuantumAnnealingAnalytics.plot_states_dwisc — Method

function to plot only the all states in a dwisc bqpjson format file, String, or Dict. possible orderings are :numeric, :hamming, or :energy. spin_comp specifies a state in spin format for comparison in hamming distance calculations. kwargs is for Plots.bar